Revert "pyrofork: Refactor StarGift and remove duplicate types"
This reverts commit 65a2febd2544d2a1f08ba9fd3af7dce09946a2b1.
